Moved Util to parent package, added setUpOutputDirectory
