+++ serializer/src/main/java/serializer/TryToParseMutants.java
@@ -4,7 +4,6 @@ import com.github.javaparser.Problem;
 import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.TokenRange;
 import com.github.javaparser.ast.CompilationUnit;
-import serializer.peg.Util;
 
 import java.io.File;
 import java.io.FileNotFoundException;
